Implementa Usecases. Implementa Repository. Configura Koin para Domain.

+++ application/data/src/main/java/com/hugothomaz/data/cloud/request/freight/FreightPriceRequest.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class FreightPriceRequest(
     @SerializedName("axis") val axis: Int,
-    @SerializedName("distance") val distance: Double,
+    @SerializedName("distance") val distance: Int,
     @SerializedName("has_return_shipment") val hasReturnShipment: Boolean
 )
