+++ application/data/src/main/java/com/hugothomaz/data/cloud/request/router/RouterRequest.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 import com.hugothomaz.data.cloud.request.router.PlaceRequest
 
 data class RouterRequest(
-    @SerializedName("fuel_consumption") val fuelConsumption: Int,
+    @SerializedName("fuel_consumption") val fuelConsumption: Double,
     @SerializedName("fuel_price") val fuelPrice: Double,
     @SerializedName("places") val placeRequests: List<PlaceRequest>
 )
